chore: update to Docusaurus v3
