+++ .gitignore
@@ -23,7 +23,7 @@ node_modules
 /versioned_docs
 /versioned_sidebars
 /versions.json
-/docusaurus.config.js
+/docusaurus.config.ts
 /playwright/
 /test-results/
 /playwright-report/
