+++ package.json
@@ -30,19 +30,19 @@
     "version-java": "npm run prepare-java && docusaurus docs:version stable && cp -r docs/* java/docs/ && cp -r versioned_docs/* java/versioned_docs/ && cp -r versioned_sidebars/* java/versioned_sidebars/ && cp versions.json java/",
     "version-node": "npm run prepare-nodejs && docusaurus docs:version stable && cp -r docs/* nodejs/docs/ && cp -r versioned_docs/* nodejs/versioned_docs/ && cp -r versioned_sidebars/* nodejs/versioned_sidebars/ && cp versions.json nodejs/",
     "version-python": "npm run prepare-python && docusaurus docs:version stable && cp -r docs/* python/docs/ && cp -r versioned_docs/* python/versioned_docs/ && cp -r versioned_sidebars/* python/versioned_sidebars/ && cp versions.json python/",
-    "cleanup": "rm -rf docs versioned_docs versioned_sidebars docusaurus.config.js versions.json sidebars.js",
+    "cleanup": "rm -rf docs versioned_docs versioned_sidebars docusaurus.config.ts versions.json sidebars.js",
     "version-all": "npm run version-python && npm run version-node && npm run version-java && npm run version-dotnet",
     "serve": "docusaurus serve",
     "build-analytics-plugin": "tsc -p plugins/playwright-analytics-integration/tsconfig.json"
   },
   "dependencies": {
-    "@docusaurus/core": "^2.4.0",
-    "@docusaurus/plugin-content-docs": "^2.4.0",
-    "@docusaurus/preset-classic": "^2.4.0",
-    "@docusaurus/types": "^2.4.0",
-    "clsx": "^1.2.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "@docusaurus/core": "3.0.0",
+    "@docusaurus/plugin-content-docs": "3.0.0",
+    "@docusaurus/preset-classic": "3.0.0",
+    "@docusaurus/types": "3.0.0",
+    "clsx": "^2.0.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-lite-youtube-embed": "^2.3.52",
     "react-loadable": "^5.5.0"
   },
@@ -59,19 +59,18 @@
     ]
   },
   "devDependencies": {
-    "@docusaurus/module-type-aliases": "^2.4.0",
-    "@playwright/test": "^1.32.0",
-    "@tsconfig/docusaurus": "^1.0.7",
-    "@types/node": "^16.18.10",
-    "@types/react": "^17.0.52",
+    "@docusaurus/module-type-aliases": "3.0.0",
+    "@docusaurus/tsconfig": "3.0.0",
+    "@playwright/test": "^1.40.0",
+    "@types/node": "^20.9.2",
+    "@types/react": "^18.2.37",
     "@types/react-router-dom": "^5.3.3",
-    "@types/rimraf": "^4.0.5",
     "chokidar": "^3.5.3",
-    "concurrently": "^7.6.0",
-    "esbuild-loader": "3.0.1",
+    "concurrently": "^8.2.2",
     "http-server": "^14.1.1",
     "lodash": "^4.17.21",
-    "rimraf": "^4.4.1",
-    "typescript": "^5.0.2"
+    "rimraf": "^5.0.5",
+    "swc-loader": "^0.2.3",
+    "typescript": "^5.2.2"
   }
 }
