+++ tsconfig.json
@@ -1,7 +1,12 @@
 {
-    "extends": "@tsconfig/docusaurus/tsconfig.json",
+    "extends": "@docusaurus/tsconfig/tsconfig.json",
+    "include": [
+        "src/"
+    ],
     "compilerOptions": {
       "baseUrl": ".",
-      "downlevelIteration": true
+      "downlevelIteration": true,
+      "moduleResolution": "Node16",
+      "module": "Node16"
     },
 }
